+++ src/toolbar/Toolbar.tsx
@@ -8,6 +8,7 @@ import { Toolbar as UtilsToolbar } from '../components/toolbar/Toolbar';
 import BluetoothButton from './buttons/bluetooth/BluetoothButton';
 import ReplButton from './buttons/repl/ReplButton';
 import RunButton from './buttons/run/RunButton';
+import SponsorButton from './buttons/sponsor/SponsorButton';
 import StopButton from './buttons/stop/StopButton';
 import TourButton from './buttons/tour/TourButton';
 
@@ -16,6 +17,7 @@ import './toolbar.scss';
 // matches ID in tour component
 const bluetoothButtonId = 'pb-toolbar-bluetooth-button';
 const runButtonId = 'pb-toolbar-run-button';
+const sponsorButtonId = 'pb-toolbar-sponsor-button';
 const tourButtonId = 'pb-toolbar-tour-button';
 
 const Toolbar: React.VFC = () => {
@@ -34,6 +36,7 @@ const Toolbar: React.VFC = () => {
                 <ReplButton id={replButtonId} />
             </ButtonGroup>
             <ButtonGroup className="pb-toolbar-group pb-align-right">
+                <SponsorButton id={sponsorButtonId} />
                 <TourButton id={tourButtonId} />
             </ButtonGroup>
         </UtilsToolbar>